Updated openstack/openstack Project: openstack-infra/system-config  26d2f1aea0141ed0ea19a3c0359e590ed3da7b32
launch-node.py: save key when failing early

Save the key to a file in /tmp when failing early with --keep.
Although it is put into the JobDir later, if we fail before that we're
locked out of the host.

While we're here, make what just happened in an error case a little
clearer

Change-Id: Ide601e2018302664bc4ad609c4483aa1451b3724