Updated openstack/openstack Project: openstack-infra/infra-specs  ab517e56ce207d34b90629f67aed51b4ef09b563
Zuulv3: drop variable interpolation and add nodesets

The only example of interpolated variables in the spec was in order
to easily specify what kind of node should be used in a single-node
job.  Actually implementing the interpolation is not trivial, and
this kind of interpolation adds some risk around making the
configuration language more complicated, difficult to follow, and
error-prone for users.

Monty suggested that perhaps nodesets could become a first-class
configuration option and negate the need for this particular example.
That seems to work quite well -- allowing sets of nodes to be defined
and referred to by name.

Without this example to stand on, the variable interpolation feature
doesn't seem worth it, at least at the moment.  This change removes
it and adds Nodesets as a configuration item.

If we need it later, we can add it back.  If we need a way to pass
job variables to ansible, we can add a 'vars' job attribute, but treat
it as a simple static structure, without interpolation.

Change-Id: If2de49a698948279fb8deb88eb82bb4ae163f6b3