Updated openstack/openstack Project: openstack/cinder  5038b0e210ec8343c08601a7d06a0a57f1c2bcee
EMC VMAX - multi-attach failure to VMAX3 when SLO is omitted

An SLO of NONE is a valid FAST configuration so when the SLO was
not included in the xml file we assigned NONE to it. The omission
of SLO should mean Non FAST in VMAX3, similar to how VMAX2 is
implemented.

Change-Id: I6cafd45d0f2584c0cc2aba427546c153ca53a101
Closes-Bug: #1567577