Updated openstack/openstack Project: openstack/oslo.db  b19738a13a08f9ff5ac2357060250d121c680163
Add additional caution looking for table, info

A joined inheritance mapper is mapped to a join();
these have no .info attribute, so use that
of the local table instead.  The sorting order when
querying against a joined subclass table
can at least be derived from the base table, so
use the base mapper's table.  It may be a better idea
to comprise the constraints amongst the base table
*and* the joined table, though.

Change-Id: I9466c9dbbbdc4af10ab0f15ee0f558199973c1ec