Updated openstack/openstack Project: openstack-infra/nodepool  adbf99a58b44d42767c2dd32d812f0f762d47f71
Don't delete images in a building state

This is a followup to an earlier change that prevented images from being
deleted from the dib image table if they were in a building state. This
prevents them from being removed from the snapshot table if they are in
a building state because this is where the dib images "go" when they are
uploaded to a cloud provider.

Change-Id: I0171495a76e26424d707b8fbfc508a403850565a