Updated openstack/openstack Project: openstack/nova  2bd800ce25009564f797af05c84d64c7edfd47f4
correctly join the usage to inventory for capacity accounting

This updates the SQL code to match the comment about how the tables
should be joined. It also adds a detailed warning when we go over
capacity, that will be useful to developers and operators when fail to
land on an environment.

The bug was deep in the SQL, and can only be tested by testing the
SQL, which we may want to do later. In leiu of a unit test, an
additional safety check is put in the code that if we ever end up with
multiple records with the same key, we explode with a KeyError.

Closes-Bug: #1620736

Change-Id: Ic3f83a881e7d98fc6ed007b62a179b3153190543