Updated openstack/openstack Project: openstack-infra/project-config  c3cb86298871fdd8da7819965fe7a9662a1fb887
Enable QoS on neutron for python-openstackclient.

Testing the sdk & client integration to neutron we need to enable
the qos service and extension.

This patch enables this by importing the on-tree neutron plugin
and enabling the q-qos service.

Neutron plugin is imported because 'q-qos' is only on the Neutron
in-tree plugin; otherwise 'q-qos' won't work.

Change-Id: I8db441111ec4fb0498bee83971caa37a4438d763
Related-bug: 1609472