Updated openstack/openstack Project: openstack-infra/project-config  a2b6a423bd51bed74041a4116af4a013612db1a9
Allow deb-packaging-core to push tags

We need to push tags to the git, otherwise, we don't have a ref
to build the orig.tar.xz files. For example, Alembic upstream uses
rel_0.6.2, which isn't nice for us. We need to push instead a new
label 0.6.2, which will be used by "git archive" to generate the
tarball.

On top of that, we need to be able to sign normal debian/newton
branch whenever we upload to Debian.

Change-Id: Ib6f9179134abfd38479096de3e99b99a7afe34a0