Updated openstack/openstack Project: openstack-infra/storyboard-webclient  64bbab44e19c8486a9b8bb7da8ece3299d69a7e5
Add a way to add tasks to worklists from the story view

This commit adds an "Add to worklist" button for tasks in the task list
of a story. This button opens a modal which allows the user to browse
for and select a worklist to add the task to.

This can be used to quickly add a task to a worklist representing your
priorities for example.

Task: 2747
Change-Id: I9a3d892d463aa87a4682da17a2dae3aca962c7a5