Updated openstack/openstack Project: openstack/tempest  0b75824303a94de2b7c9ccdc0a982c40b14504ce
Remove base_types_client for the reability

The service clients for the volume service are different from the
other services and it is a little difficult to read and debug because
that consists of multiple layers for v1 and v2.
That could be useful for the code optimization, but actually the merit
seems small and unreadable.
This patch removes these layers from the volume clients.

Partially implements blueprint consistent-service-method-names

Change-Id: I0d59b16d706ca3ca0bf629b1374530cc5927b7c2