Updated openstack/openstack Project: openstack-infra/puppet-mediawiki  43b131e06d1d01813abc741b839cca187abf1f94
Conditionally define the default extension source

Since $type and $source are both class parameters, we can't
interpolate $type in the default definition for $source. Instead
redefine $source inside the resource if it's unchanged from the
resource default.

Also make the contents of /src/mediawiki/w require the vcsrepo
resource for it. Early creation of the directory causes vcsrepo to
fail to clone into it.

Change-Id: I9b447173d4966f03a195825047cf25afc1a4afd6