Updated openstack/openstack Project: openstack/keystone  e91c6fb7dc397c0d6d64769d183b9fd78124f9a8
Keep the order of passwords in tests

While multiple passwords are stored for a user, only the latest one is
used for authentication. When `created_at` in tests get changed, the
order of password might change too.

Keep the order of passwords by giving earlier passwords an earlier date
of creation.

Change-Id: I68861bc5c379dcf1cf060081f7802917d6f0c4b9