Updated openstack/openstack Project: openstack/cinder  041b6f7a52fdca2ee5593035f18fb1c883fdb7fe
Mark Scality volume driver as unsupported

The Scality CI has not reported on a patch since 2016-04-12 09:44:38.
This is well past our policy for CI reporting.

Per Cinder's non-compliance policy [1] this patch marks the driver
as unsupported and deprecated and it will be approved if the issue
is not corrected by the next cycle.

[1] https://wiki.openstack.org/wiki/Cinder/tested-3rdParty-drivers#Non-Compliance_Policy

Change-Id: Ic749db69d793359c9328c8c62b919197c1cec859