Updated openstack/openstack Project: openstack/horizon  9e711d3e11aa0d4df4cd2afbf5c71de92fcc25af
Add message for a result of update metadata

At the moment, even when a user update a metadata, this result is
not informed to a user (any message is not displayed).
So a user cannot know its result and he has to redisplay
'Update Metadata' modal in order to know its result.

Change-Id: I02b1a3c70d0225d2326d038b7519d49b0082985e
Closes-Bug: #1608361