Updated openstack/openstack Project: openstack/ironic  749ed035a09d624ae3ddb1b44cf24c3823dd3197
Log a warning for unsupported drivers and interfaces

This iterates over all loaded drivers/interfaces at factory
initialization time, and looks at an attribute 'supported'. If supported
is not True, a message will be logged that the driver is unsupported,
deprecated, and may be removed in a future release.

This patch marks all drivers as supported, and subsequent patches will
mark specific drivers as unsupported, based on their testing status.

This also changes the NetworkInterface class to inherit from
BaseInterface (to make this work right). That looks like a programming
error when that interface was created.

Change-Id: I6f9fa4f367bd2e14e5992d2e601bfdcf852ef3a4
Partial-Bug: #1526410