Updated openstack/openstack Project: openstack/neutron  717bed95fe731cd9877bb63691f022b012f66b9a
Don't return ext net DB object in auto allocate

Returning a DB object for the external network and passing
that to the create_router operation makes the life of the
DB object transition many operations. If any of those operations
detech it from the session or expire it, it can result in
an exception down in the router operations.

This patch just returns the network ID since that's all we need.

Closes-Bug: #1616126
Change-Id: Ib9c4ba9f00ca7f7a713a89be4f75a230b086ac38