Updated openstack/openstack Project: openstack-infra/nodepool  5b058abc4fc702afaf2bab73b95830205219b80f
Disable searching for discoverable private keys

Paramiko has the ability to search ~/.ssh and ~/ssh for private keys.
Since we already define our private keys, there is no reason to try to
find more.

Since the _auth() method in paramiko attempts multiple forms of
authentication with the remote host, we are disabling this support to
better control the code path that paramiko takes if our defined
private keys failed.

[1] https://github.com/paramiko/paramiko/blob/v1.17.2/paramiko/client.py#L470

Change-Id: I07401818d33b2fbc6c773c0e0686fbd51610067c
Signed-off-by: Paul Belanger <pabelanger@redhat.com>