Updated openstack/openstack Project: openstack/nova  b61cb28e9837fa2820ad70d78d3a109a5f0f07c6
Change api-site to v2.1 format

The endpoint of v2.1 API is CURRENT now.
On v2.1 API, a tenant-id is not necessary to be specified on the
URL, that is good for making API calls simple for users.
In addition, the endpoint is different from old v2.0 API.
Then this patch makes the api-site fit for v2.1 format.

Partially implements blueprint api-ref-in-rst

Change-Id: I99224dfa77d64f7800733c1af69a55b3b5c761ac