Updated openstack/openstack Project: openstack/neutron  1588559407a849707b8bd0f84d2d64cfbb110769
gate_hook: Compile ovs only with supported kernels

The intention of this patch is not to break current compilations on
Xenial boxes that run with 4.4 kernel which is not supported by OVS
openvswitch kernel module. After we are confident to switch master
branch to be running on Xenial only, we will get rid of the compilation
itself until it's needed by other features.

Change-Id: I1c890bf5a6564a69e36b8289a5d8dc0deee2a429