Updated openstack/openstack Project: openstack/cinder  8753a3401b28c7d9dab644b0743a9f5d9abd602a
Fix invalid services caused by enabled_backends

When the value of enabled_backends in cinder.conf starts with a comma,
splitting the string to backends does not filter the empty backend at
the beginning and the result is an invalid cinder-volume service that
fails to start. The same applies to enabled_backends ending with comma
or having extra commas in the middle.

Change-Id: Ib18d109d28d97e2d3681eff3784f08fd82bfee71
Closes-Bug: #1611353