Updated openstack/openstack Project: openstack-infra/project-config  9560cbdf8d3b8540e3f12d3f760e151e3bd022d0
Revert "Remove deb-websockify from projects.yaml"

deb-websockify was broken, and therefore removed. Since, the upstream
repo on alioth.debian.org [1] was fixed as per recommendation from the
infra team. Clarkb asked for this commit to be reverted, so here we are.

Note: a gerrit admin will have to force push the repo contents

[1] https://anonscm.debian.org/git/openstack/deb-websockify.git

This reverts commit dbe0206ceb6376a85f105af0658327a53f3992cf.

Change-Id: Ia414e8bab6ad613b43d4c2626df424fee157740a