Updated openstack/openstack Project: openstack/nova  c613b083a927c54080f74005a4a797aea9c91d98
db: retry on deadlocks while adding an instance

We are hitting deadlocks in the gate when we are inserting the new
instance_extra row into the DB.

We should follow up this fix and look at way to avoid the deadlock
happening rather than retrying it. It currently doesn't happen too
often, so this should be enough to stop the problem while we work on a
better fix.

Closes-Bug: #1480305

Change-Id: Iba218bf28c7d1e6040c551fe836d6fa5e5e45f4d