Updated openstack/openstack Project: openstack/python-openstackclient  7d1a5d0854c732681a130255ddc6abb2e9721a7a
format token expires time to prevent json loading datetime data

When output to shell, the token issue time is fine; however when
selecting the json formatter (via the --format json) option, an
exception is raised when formatting the dataetime data.

Rather than pass in the datetime data, we should format the data
with the ISO 8601 formatting.

Closes-Bug: 1619937
Change-Id: Iffebb2d5413fabfd283dfa94fc560fc37270f9dd