Updated openstack/openstack Project: openstack-dev/devstack  95469032656c442982730ea307644d47d7755a81
Set sensible default for LB_PHYSICAL_INT

Configure the linux bridge physical interface to use the interface for
the default route on the current host. In the future we should consider
using a dangling interface so that we aren't affecting the host
instances networking but this roughly matches what testing has been
using in the past.

Change-Id: I7859437f97e6cab929e90208fe56f7efd62dfe01