Updated openstack/openstack Project: openstack/tripleo-heat-templates  0f3249df2b68ff426a6aff6ccd8981747e234586
Add Ceph cluster health validation on upgrade

This will prevent the Ceph Mon upgrade script from starting if the
Ceph cluster is in error state.

It also adds a parameter to ignore warning states, useful when
performing an upgrade of a cluster where the number of healthy
OSDs does not guarantee the desired replica size.

Closes-Bug: 1618533
Change-Id: I1beb8ad0812f19b1018ba19b5a9fc85fa132d7f7