Updated openstack/openstack Project: openstack-infra/project-config  a060d48e5d72ec979019a2ca4771eedc7a8df2c9
setup post job for sahara-extra

this changes setups additional post job for sahara-extra artifacts.
we already have the script that make all needed build operations, and
it lives in sahara-extra repository.

since some these artifacts should be stored per branch, we are passing
branch as argument of the script so that after successful build all
needed stuff will be in to appropriate place ('dist/newton/' directory
for example).

Change-Id: I2d2a29f71ba11da46c9a313132af39709cc1ef61