Updated openstack/openstack Project: openstack/neutron  21dbf03e25995377955257204a16a90e60b335c0
objects: remove support for multiple db models in from_db_object

This was once implemented to allow objects that aggregate attributes
from multiple models. Specifically, that was the original design for QoS
rules, where we had a model with common attributes, and then a separate
model with rule specific attributes.

After a while, struggling to make it work without huge hacks, and after
rethinking database schema, we dropped the original design for QoS
rules, replacing it with a single model per type.

This patch cleans up the remnants of those dark ages of QoS.

This should make the code a bit more clear to readers who currently
wonder why we may have multiple database models for the same versioned
object.

Change-Id: I6a083f95ec02f61ced8f7ad0117560347d64b369
Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db