Updated openstack/openstack Project: openstack-infra/nodepool  59ff94f947b1da168d9f9c691b36294774a81594
Disable connecting to the SSH agent for ssh_connect

To avoid some of the magic that paramiko does, we'll be disabling
support for the SSH agent. This is intended to limit the code path in
paramiko when debuging SSH failures.  Additionally, nodepool already
defines the path to SSH private keys no need to try others.

Change-Id: I4d8350089187dfb78932766b8a5a37cffd9385f5
Signed-off-by: Paul Belanger <pabelanger@redhat.com>