Updated openstack/openstack Project: openstack/ironic  af40281b5aa1ece4f255d1f316934d2fe429ecb0
Add a basic install guide

This installs the basic directories and content pointing to the current
Ironic Installation Guide. Simply having this in our repo will not lead
to it being published, this is only the first step. Having these items
in our repository is a good first step to building out our install
guide.

Change-Id: I6a0e69bd3b6f5074ae61505f82f27bcf9df84bab
Partial-bug: #1612278