Updated openstack/openstack Project: openstack/nova  53216e47fc906b2f657c35b7f1e6bd2f364eb6d1
Ensure ResourceProvider/Inventory created before add Allocations record

In the startup of compute node, the first available resource update
will fail to add Allocations record. This is due to ResourceProvider
will be created at the end of available resource update.

This patch ensure the ResourceProvider and Inventory created in the
beginning of available resource update.

Closes-Bug: #1621437
Change-Id: I91fa7b8354a7613848225429ac9d38c6305b03c6