Updated openstack/openstack Project: openstack/tripleo-heat-templates  3cc34744f1b116b16142caa31d7a3537420dd70f
Generate VIP info for ctlplane VIP, not management

The management network does not have a VIP, so it's been wrong to
generate a cloud name and hieradata for this. Instead, the network
that actually needs a name and a hosts entry is the ctlplane network,
which actually has a VIP and there are services that use it.

bp tls-via-certmonger

Closes-Bug: #1621742
Change-Id: I163b2c7b5684da6dc290636f54eefe3f2b0c3e3f