Updated openstack/openstack Project: openstack/keystone  179c5605f97419a9b573cebaa0844ec276519df3
Use freezegun for change password tests

We recently added a test for PCI password compliance that changes
passwords and attempts to get a token right afterwords. This causes
race conditions with Fernet tokens and we typcially use freezegun in
the tests to address it.

This commit does follows the same convention we've done elsewhere in
keystone.

Change-Id: I34e60b6fe56ae32ae0a32c8345b951bf9fe53ba9