Updated openstack/openstack Project: openstack/nova  d8d755f1b1f6857991fc23fbfe9021b6cc0870c9
Fix resizing in imagebackend.cache()

The Flat and Lvm backends do not create a 'base image' (the file in the
image cache) when creating an ephemeral or swap disk. However, cache()
expects it to exist when checking if a resize is required.

This change ignores the resize check if the backing file doesn't exist.
This happens to be ok, because ephemeral and swap disks are always
created with the correct target size anyway, and therefore never need
to be resized.

Closes-Bug: 1608934
Co-Authored-By: Matthew Booth <mbooth@redhat.com>
Change-Id: I46b5658efafe558dd6b28c9910fb8fde830adec0