Updated openstack/openstack Project: openstack/tempest  c3fef9f530ea4225468802bbf40e0b4d780aa11d
Switch unit test tox jobs to use ostestr

Now that we've deprecated all our bash wrappers for running tests
we shouldn't be using them for running unit tests either. This commit
updates the unit test tox job definitions so that we no longer are
relying on deprecated code in the gate and we can remove the wrappers
when the time is right.

Change-Id: I127786537c0792456d26b13fc3a2044afc4e93a1