Updated openstack/openstack Project: openstack/keystone  bf5711ef029e230fb8d9daae34f0528c8b68a9ae
Remove the dead link in schema migration doc

The schema migration document refers to a another document that
doesn't exist. This patch removes the reference since no appropriate
link could be found.

The doc was also updated to include the copyright.

This fix also resolves a warning in the tox doc build:

/keystone/doc/source/online_schema_migration_examples.rst:23:
WARNING: undefined label: online-migration (if the link has no
caption the label must precede a section header)

Change-Id: I4732651f8c5efd9fa494fd5fba5dad1f782d89a2
Closes-Bug: #1622035