Updated openstack/openstack Project: openstack/nova  085ddd7bbf6633955e76634613cec5e70594776d
virt: handle unicode when logging LifecycleEvents

The repr on the LifecycleEvent object includes a translated
name, which blows up with a UnicodeEncodeError in the emit_event
method because of str(event) and non-English locales.

This change uses six.text_type on the object rather than str
and adds a test to recreate the bug and show the fix.

Change-Id: I9b7b52739883043b7aae9759f500e5e21cfe8b30
Closes-Bug: #1621392