Updated openstack/openstack Project: openstack/manila  eddeff89bd48b8180c5907ea96283658d4b21a46
Check 'thin_provisioning' in extra specs

Currently in the capacity filter and weigher of the scheduler,
we use the logic to evaluate whether there is enough capacity
to thin provision a share on a backend if the driver reports
thin_provisioning to be True. However, a driver may be able to
support both thin and thick provisioning. The logic does not
check whether the user wants the share to be provisioned as thin
or not.

In this patch, we check 'thin_provisioning' in extra specs of the
share type and decide whether to use the logic for thin or thick.
In the following two cases, we will use the thin logic:
1) 'thin_provisioning' is not set in extra specs. This is to
   provide backward compatibility.
2) 'thin_provisioning' is set in extra specs and it is
   '<is> True' or 'True'.

To provision a thick share on a backend that supports both
thin and thick, set one of the following in extra specs:
    {'thin_provisioning': 'False'}
    {'thin_provisioning': '<is> False'}
    {'capabilities:thin_provisioning': 'False'}
    {'capabilities:thin_provisioning': '<is> False'}

DocImpact
Change-Id: I238a7962425ea35c356c5ed2e31b8f68462b3769
Closes-Bug: #1578718