Updated openstack/openstack Project: openstack-infra/project-config  8bb2903ee27c6d0d6ae6556de9bb1550df07f295
Add new project networking-zte

Create a seperate project for Neutron integration
with ZTE networking. The logic of communicating
with ZTE controller(ZENIC) and ZTE Hardware device
is contained in this project. It implements ZTE ML2
mechanism driver and ZTE neutron agents for Openstack
Neutron.

It will start with an empty repository and review with
voting jobs.

Change-Id: I128bd8e4090617a2df6900388c4bf6739bba1633